split config.cc into config.cc and config.h
